refactor turn Currency into Price
